soc2_backup/backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
groq==0.15.0
pypdf2==3.0.1
python-dotenv==1.0.1
pydantic==2.10.4